requirements.txt
datedelta~=1.3
requests~=2.25.1
pandas~=1.2.2
xlwings~=0.22.2
pytest~=6.2.2
setuptools~=53.0.0
openpyxl
mysql-connector-python